kobayashi_maru v0.0.1 KobayashiMaru.Accounts.User

I have added a password virtual field, to temporarily hold the clear-text password. This will never be saved to the DB. I also added two changesets, one for changing the password and one for the other fields. The put_password_hash() function uses the comeonin library to get a hash of the password and put that on the password_hash field that will be stored in DB.

Functions

changeset(user, attrs)

changeset(
  KobayashiMaru.Accounts.User.t(),
  :invalid
  | %{optional(:__struct__) => none(), optional(atom() | binary()) => any()}
) :: Ecto.Changeset.t()

registration_changeset(user, attrs)

registration_changeset(
  KobayashiMaru.Accounts.User.t(),
  :invalid
  | %{optional(:__struct__) => none(), optional(atom() | binary()) => any()}
) :: Ecto.Changeset.t()